Update README.md
changed how I refer to what I will post about my game. I still don't know if these are the right words for it. if anybody knows what the word is for a pre release to test part of the thing is then that's the word I'm grasping for here. maybe beta is the right word but it's not the full game? will finish crossing that bridge when I come to it.